Extract `InstantExecutionCacheFingerprint` sealed type hierarchy to a separate file
