+++ subprojects/instant-execution/src/main/kotlin/org/gradle/instantexecution/extensions/VirtualFileSystemExtensions.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.instantexecution.extensions
+
+import org.gradle.internal.hash.HashCode
+import org.gradle.internal.vfs.VirtualFileSystem
+import java.io.File
+
+
+internal
+fun VirtualFileSystem.hashCodeOf(file: File): HashCode? =
+    readRegularFileContentHash(file.path) { hashCode -> hashCode }
+        .orElse(null)
